+++ src/main/java/com/space/service/ShipService.java
@@ -9,6 +9,6 @@ public interface ShipService {
     void add(Ship ship);
     void delete(Ship ship);
     void edit(Ship ship);
-    Ship getById(int id);
+    Ship getById(Long id);
     int shipsCount();
 }
